requirements.txt
langchain-community 
# langchainhub 
langgraph
langchain[groq]
streamlit
python-dotenv